+++ api/cakes/getPgTokenData.ts
@@ -1,0 +1,8 @@
+import PATH from '../common/path';
+import { client } from '../common/axios';
+
+export const getPgTokenData = async (pgToken: string) => {
+  const data = await client.get(`${PATH.CAKES}/${PATH.APPROVE}?${PATH.PG_TOKEN}=${pgToken}`);
+
+  return data;
+};
