+++ api/cakes/getWishesData.ts
@@ -1,10 +1,8 @@
-import axios from 'axios';
 import PATH from '../common/path';
+import { client } from '../common/axios';
 
 export const getWishesData = async (wishesNumber: number) => {
-  const data = await axios.get(
-    `https://www.sunmulzu.shop/${PATH.API}/${PATH.V1}/${PATH.WISHES}/${wishesNumber}`,
-  );
+  const data = await client.get(`${PATH.API}/${PATH.V1}/${PATH.WISHES}/${wishesNumber}`);
 
   return data.data.data;
 };
